try to trigger redeploy with a squash
