+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@lob/ui-components",
-  "version": "1.0.0-beta.43",
+  "version": "1.0.0-beta.44",
   "engines": {
     "node": ">=14.18.2",
     "npm": ">=8.3.0"
