+++ src/components/Button/Button.stories.js
@@ -1,5 +1,6 @@
 import LobButton from './Button.vue';
 import mdx from './Button.mdx';
+import { Plus, Download } from '@/components/Icons';
 
 export default {
   title: 'Components/Button',
@@ -16,22 +17,26 @@ export default {
       },
       description: 'Content to display inside of the button',
       table: {
-        defaultValue: 'Sign up',
+        defaultValue: 'This is a button',
         type: {
           summary: 'html or component'
         }
       }
     },
     variant: {
-      options: ['primary', 'secondary', 'tertiary', 'alert', 'success', 'error', 'link', 'none'],
+      options: ['primary', 'secondary', 'link', 'none'],
       control: {
         type: 'select'
       }
     },
-    size: {
-      options: ['default', 'small', 'large'],
+    warning: {
       control: {
-        type: 'select'
+        type: 'boolean'
+      }
+    },
+    small: {
+      control: {
+        type: 'boolean'
       }
     }
   }
@@ -44,7 +49,23 @@ const Template = (args, { argTypes }) => ({
   template: '<lob-button v-bind="args" v-slot>{{ args.content }}</lob-button>'
 });
 
-export const Primary = Template.bind({});
-Primary.args = {
-  content: 'Sign up'
+export const Default = Template.bind({});
+Default.args = {
+  content: 'This is a button'
+};
+
+const WithIconsTemplate = (args, { argTypes }) => ({
+  props: Object.keys(argTypes),
+  components: { LobButton, Download, Plus },
+  setup: () => ({ args }),
+  template: `<lob-button v-bind="args"> 
+              <Download class="w-6 h-6" style="margin-right:10px;"/>
+              {{ args.content }} 
+              <Plus class="w-6 h-6" style="margin-left:10px;"/>
+            </lob-button>`
+});
+
+export const WithIcons = WithIconsTemplate.bind({});
+WithIcons.args = {
+  content: 'This is a button'
 };
