+++ src/components/Link/Link.vue
@@ -1,50 +1,64 @@
 <template>
-  <component
-    :is="tag"
-    :[linkProp]="to"
-    :target="target"
-    :rel="target === '_blank' ? 'noopener noreferrer' : ''"
-    :aria-disabled="disabled"
-    :class="[
-      {'font-bold': bold},
-      {'underline' : underline},
-      {'text-primary-900': !inheritTextColor },
-      {'text-inherit': inheritTextColor },
-      {'primary py-3 px-6 bg-primary-500 !text-white active:bg-primary-700': primary},
-      {'secondary py-3 px-6 bg-white-200 border border-primary-500 text-primary-500 active:text-primary-700 active:border-primary-700': secondary},
-      {'tertiary py-3 px-6 bg-white border border-gray-100 text-gray-500 active:border-gray-300': tertiary},
-      {'alert py-3 px-6 bg-white hover:bg-lemon-300 border rounded border-lemon-700 text-lemon-900 hover:text-lemon-900 focus:ring-lemon-700': alert},
-      {'flex justify-center items-center rounded focus:outline-none focus:ring-2 focus:ring-primary-100 focus:border-transparent': primary || secondary || tertiary},
-      {'hover:text-white': primary},
-      {'!px-3 !py-2': small},
-      {'disabled pointer-events-none !text-gray-500': disabled},
-      {'bg-white-300': disabled && primary},
-      {'border-gray-100': disabled && secondary},
-      {'border-white-300': disabled && tertiary},
-      {'border-gray-100 hover:bg-white': disabled && alert}
-    ]"
-  >
-    <slot />
-  </component>
+  <span :class="[{ 'cursor-not-allowed': disabled, 'cursor-pointer': !disabled }]">
+    <component
+      :is="tag"
+      :[linkProp]="to"
+      :target="target"
+      :rel="target === '_blank' ? 'noopener noreferrer' : ''"
+      :aria-disabled="disabled"
+      :class="[
+        { 'pointer-events-none': disabled },
+        { 'font-bold': bold },
+        { 'text-inherit': inheritTextColor },
+        { 'px-0 text-base focus-visible:ring-4 focus-visible:ring-offset-2 focus-visible:rounded-lg': link },
+        { 'underline': underline && link },
+        { '!text-gray-500': disabled && link },
+        { 'text-primary-500 hover:text-primary-900 active:text-primary-900': !inheritTextColor && link },
+        { 'flex justify-center items-center rounded-lg no-underline font-medium': primary || secondary },
+        { 'focus-visible:ring-4 focus-visible:ring-primary-100 focus:ring-transparent focus:outline-none active:scale-[.96]': primary || secondary },
+        { 'px-6 text-base h-[48px]': regular && (primary || secondary) },
+        { 'px-4 text-sm h-[32px]': small && (primary || secondary) },
+        { 'primary !text-white ': !disabled && primary && !warning },
+        { 'bg-gray-100 !text-white': disabled && primary && !warning },
+        { 'primary warning !text-white': !disabled && primary && warning },
+        { 'bg-coral-200 !text-white': disabled && primary && warning },
+        { 'secondary border !border-gray-300 text-gray-500 hover:text-gray-500 hover:bg-gray-100/[.15] active:bg-gray-100/[.25]': !disabled && secondary && !warning },
+        { 'border border-gray-100 text-gray-100': disabled && secondary && !warning },
+        { 'secondary bg-white border !border-chili text-chili hover:bg-chili/[.04] active:bg-chili/[.08]': !disabled && secondary && warning },
+        { 'border border-coral-200 text-coral-200': disabled && secondary && warning }
+      ]"
+      v-bind="$attrs"
+    >
+      <slot />
+      <ChevronRight
+        v-if="link && withChevron"
+        class="h-4 inline-flex -mr-1"
+      />
+    </component>
+  </span>
 </template>
 
 <script>
+import { ChevronRight } from '@/components/Icons';
 export default {
   name: 'LobLink',
+  components: { ChevronRight },
+  inheritAttrs: false,
   props: {
     variant: {
       type: String,
-      default: 'default',
+      default: 'link',
       validator: function (value) {
-        return ['default', 'primary-button', 'secondary-button', 'tertiary-button', 'alert-button'].includes(value);
+        return ['link', 'primary-button', 'secondary-button'].includes(value);
       }
     },
-    size: {
-      type: String,
-      default: 'default',
-      validator: function (value) {
-        return ['default', 'small'].includes(value);
-      }
+    small: {
+      type: Boolean,
+      default: false
+    },
+    warning: {
+      type: Boolean,
+      default: false
     },
     disabled: {
       type: Boolean,
@@ -73,11 +87,15 @@ export default {
     target: {
       type: String,
       default: '_self'
+    },
+    withChevron: {
+      type: Boolean,
+      default: false
     }
   },
   computed: {
-    default () {
-      return this.variant === 'default';
+    link () {
+      return this.variant === 'link';
     },
     primary () {
       return this.variant === 'primary-button';
@@ -85,25 +103,15 @@ export default {
     secondary () {
       return this.variant === 'secondary-button';
     },
-    tertiary () {
-      return this.variant === 'tertiary-button';
-    },
-    alert () {
-      return this.variant === 'alert-button';
-    },
-    small () {
-      return this.size === 'small';
+    regular () {
+      return !this.small;
     },
     isExternal () {
       const protocolRelativePattern = /^https?:\/\/|^\/\//i;
       return typeof this.to === 'string' && protocolRelativePattern.test(this.to);
     },
     tag () {
-      if (this.isExternal) {
-        return 'a';
-      } else {
-        return 'router-link';
-      }
+      return this.isExternal ? 'a' : 'router-link';
     },
     linkProp () {
       return this.isExternal ? 'href' : 'to';
@@ -113,15 +121,21 @@ export default {
 </script>
 
 <style scoped lang="scss">
-.primary:hover:not(.disabled):not(:focus) {
-  box-shadow: 0 0 10px 2px rgba(24, 118, 219, 0.6);
+.primary:not(:disabled) {
+  background: linear-gradient(114.08deg, #1876db 7.95%, #5748ff 90.87%);
 }
 
-.secondary:hover:not(.disabled):not(:focus) {
-  box-shadow: 0 0 10px rgba(65, 101, 129, 0.2);
+.primary:focus:not(:disabled),
+.primary:hover:not(:disabled):not(:focus) {
+  background: linear-gradient(114.08deg, #5748ff 7.95%, #1876db 90.87%);
 }
 
-.tertiary:hover:not(.disabled):not(:focus) {
-  box-shadow: 0 0 10px 2px rgba(0, 153, 215, 0.2);
+.primary.warning:not(:disabled) {
+  background: linear-gradient(114.08deg, #db1818 7.95%, #ec4949 90.87%);
+}
+
+.primary.warning:focus:not(:disabled),
+.primary.warning:hover:not(:disabled):not(:focus) {
+  background: linear-gradient(114.08deg, #ec4949 7.95%, #db1818 90.87%);
 }
 </style>
