--- src/components/NewButton/NewButton.mdx
@@ -1,43 +1,0 @@
-import {
-  Preview,
-  Story,
-  ArgsTable,
-  PRIMARY_STORY,
-} from "@storybook/addon-docs";
-import NewButton from "./NewButton.vue";
-
-# Button
-
-Buttons are styled semantic buttons that will handle all of the appropriate button states.
-
-<Preview>
-  <Story id="components-newbutton--default" />
-</Preview>
-
-## How to Use
-
-There are 2 variants of buttons: 'primary' & 'secondary'; as well as 'link' and 'none' styles.
-
-You can change the variant of the button by adding the `variant` prop (String) to the component.
-
-The Button can be regular size (default, height: 48px) & or small (32px) with the `small` prop (Boolean).
-
-The `warning` prop (Boolean) will change the button color to dark red if primary, and red outline if secondary.
-
-Please note: though you can only insert text in the Storybook demo, the button component will accept any valid HTML or Vue component, as it uses a slot.
-
-Example of using this component as a primary button in a template
-
-```html
-<new-button> This is a primary button </new-button>
-```
-
-Example of using this component as a secondary button in a template
-
-```html
-<new-button variant="secondary"> This is a secondary button </new-button>
-```
-
-## Props
-
-<ArgsTable story={PRIMARY_STORY} />
