--- src/components/NewLink/NewLink.mdx
@@ -1,51 +1,0 @@
-import {
-  Preview,
-  Story,
-  ArgsTable,
-  PRIMARY_STORY,
-} from "@storybook/addon-docs";
-import NewLink from "./NewLink.vue";
-
-# Link
-
-Links are styled semantic links. If you pass through a relative path (e.g. `/settings`), the link will become a router-link integrated with Vue router. If you pass through an external URL, the link will become an anchor tag.
-
-<Preview>
-  <Story id="components-newlink--regular-link" />
-</Preview>
-
-## How to Use
-
-There are 3 variants of links: `link` (default), `primary-button` & `secondary-button`.
-You can change the variant of the button by adding the `variant` prop (String) to the component.
-The primary-button & secondary-button variants are styled so that they look like the primary & secondary Buttons.
-
-There are 2 sizes of buttons: regular (default) and small. For a small button add the `small` prop (Boolean) to the component.
-
-There is an underline property for the regular link text which is enabled by default. To disable it add the prop `:underline:"false"`.
-
-The `disabled` property applies disabled styles as well as prevents clicks via CSS.
-
-Please note: though you can only insert text in the Storybook demo, the link component will accept any valid HTML or Vue component.
-
-Example of using this component as a default link in a template
-
-```html
-<new-link to="/internal"> Click me </new-link>
-```
-
-Example of using this component as a secondary-button, small link in a template
-
-```html
-<new-link variant="secondary-button" small to="/internal"> Click me </new-link>
-```
-
-Using the component with `target="_blank"` will automatically add `rel="noopener noreferrer"` to the link
-
-```html
-<new-link target="_blank"> Click me </new-link>
-```
-
-## Props
-
-<ArgsTable story={PRIMARY_STORY} />
