+++ src/components/Table/Table.stories.js
@@ -2,13 +2,13 @@ import LobTable from './Table.vue';
 import TableHeader from './TableHeader.vue';
 import TableBody from './TableBody.vue';
 import TableRow from './TableRow.vue';
-import NewButton from '../NewButton/NewButton.vue';
+import LobButton from '../Button/Button.vue';
 import mdx from './Table.mdx';
 
 export default {
   title: 'Components/Table',
   component: LobTable,
-  subcomponents: { TableHeader, TableRow, TableBody, NewButton },
+  subcomponents: { TableHeader, TableRow, TableBody, LobButton },
   decorators: [
     () => ({ template: '<div style="width: 700px"><story /></div>' })
   ],
@@ -104,7 +104,7 @@ TableWithASingleCellRow.args = {
 
 const TableWithActionTemplate = (args, { argTypes }) => ({
   props: Object.keys(argTypes),
-  components: { LobTable, TableHeader, TableRow, TableBody, NewButton },
+  components: { LobTable, TableHeader, TableRow, TableBody, LobButton },
   setup: () => ({ args }),
   template: `
     <lob-table class="min-w-full divide-y divide-gray-200" :space="args.space">
@@ -117,7 +117,7 @@ const TableWithActionTemplate = (args, { argTypes }) => ({
         <TableRow v-for="item in args.items">
           <div class="whitespace-nowrap">{{item.name}}</div>
           <div class="whitespace-nowrap">{{item.description}}</div>
-          <div class="text-right"><new-button small>Edit</new-button></div>
+          <div class="text-right"><lob-button small>Edit</lob-button></div>
         </TableRow>
       </TableBody>
     </lob-table>
