--- src/components/Typography/P.vue
@@ -1,16 +1,0 @@
-<script>
-import { h } from 'vue';
-import Typography from './Typography';
-
-export default {
-  name: 'P',
-  components: {
-    Typography
-  },
-  render () {
-    return h(Typography, { variant: 'body1' }, { default: this.$slots.default });
-  }
-};
-
-</script>
-
