--- src/components/Typography/index.js
@@ -1,6 +1,0 @@
-export { default as Typography } from './Typography';
-export { default as H1 } from './H1';
-export { default as H2 } from './H2';
-export { default as H3 } from './H3';
-export { default as P } from './P';
-export { default as Caption } from './Caption';
