+++ src/components/index.js
@@ -3,7 +3,6 @@ export { default as Alert } from './Alert/Alert';
 export { default as Badge } from './Badge/Badge';
 export { default as Breadcrumb } from './Breadcrumb/Breadcrumb';
 export { default as LobButton } from './Button/Button';
-export { default as NewButton } from './NewButton/NewButton';
 export { default as Card } from './Card/Card';
 export { default as Calendar } from './Calendar/Calendar';
 export { default as Checkbox } from './Checkbox/Checkbox';
@@ -15,7 +14,6 @@ export { default as FilterContent } from './FilterContent/FilterContent';
 export * from './Icons';
 export { default as LoadingIndicator } from './LoadingIndicator/LoadingIndicator';
 export { default as LobLink } from './Link/Link';
-export { default as NewLink } from './NewLink/NewLink';
 export { default as LobLabel } from './LobLabel/LobLabel';
 export { default as MainNavigation } from './MainNavigation/MainNavigation';
 export { default as MainNavigationItem } from './MainNavigation/MainNavigationItem';
@@ -48,4 +46,3 @@ export { default as TopNavbar } from './TopNavbar/TopNavbar';
 export { default as Stepper } from './Stepper/Stepper';
 export { default as StepperItem } from './Stepper/StepperItem';
 export { default as StepperVertical } from './StepperVertical/StepperVertical';
-export * from './Typography';
